website/app/requirements.txt
absl-py==0.7.0
alembic==1.0.7
appnope==0.1.0
astor==0.7.1
backcall==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.2
cachetools==3.0.0
cassandra-driver==3.16.0
certifi==2017.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.7.0
cobra==0.14.2
configparser==3.5.0
cycler==0.10.0
dask==1.1.0
decorator==4.1.2
defusedxml==0.5.0
depinfo==1.5.1
dominate==2.3.5
dtw==1.3
entrypoints==0.2.3
enum34==1.1.6
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.4.0
Flask-SocketIO==3.3.1
Flask-SQLAlchemy==2.3.2
future==0.16.0
gast==0.2.2
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
grpcio==1.18.0
h5py==2.9.0
html5lib==1.0.1
httplib2==0.12.0
idna==2.8
imutils==0.5.2
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.3
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.2.2
numpy==1.15.4
oauth2client==3.0.0
oauthlib==3.0.0
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
optlang==1.4.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pathlib2==2.3.0
patsy==0.5.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.4.1
pipdeptree==0.13.2
prometheus-client==0.5.0
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-editor==1.0.4
python-engineio==3.4.3
python-libsbml==5.16.0
python-socketio==3.1.2
pytz==2017.3
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.3.1
redis==3.0.1
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.0
ruamel.yaml==0.15.88
scandir==1.6
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.18
statsmodels==0.9.0
swiglpk==4.65.0
sympy==1.1.1
tabulate==0.8.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==4.5.2
traitlets==4.3.2
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.0
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.8
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
